修正由于没有处理 std::is_trivially_copyable() 类型元素(例如 std::string)，导致内存异常 free() 的问题

+++ src/test_nut/numeric/test_biginteger.cpp
@@ -126,10 +126,10 @@ class TestBigInteger : public TestFixture
     // 测试平方
     void test_square()
     {
-        // 平法优化算法处理负数时可能出错
-        int a[2] = {-2, -1};
-        int x[6];
-        signed_multiply<int>(a, 2, a, 2, x, 6);
+        // 平方优化算法处理负数时可能出错
+        unsigned a[2] = {(unsigned)-2, (unsigned)-1};
+        unsigned x[6];
+        signed_multiply<unsigned>(a, 2, a, 2, x, 6);
         //cout << endl << x[0] << ' ' << x[1] << ' ' << x[2] << ' ' << x[3] <<
         //        ' ' << x[4] << ' ' << x[5] << endl;
         NUT_TA(x[0] == 4 && x[1] == 0 && x[2] == 0 && x[3] == 0 && x[4] == 0 && x[5] == 0);
